Add new videos section to YIR page
